+++ src/frontend/apps/e2e/__tests__/app-impress/doc-editor.spec.ts
@@ -338,6 +338,7 @@ test.describe('Doc Editor', () => {
   ].forEach(({ ai_transform, ai_translate }) => {
     test(`it checks AI buttons when can transform is at "${ai_transform}" and can translate is at "${ai_translate}"`, async ({
       page,
+      browserName,
     }) => {
       await mockedDocument(page, {
         accesses: [
@@ -364,11 +365,17 @@ test.describe('Doc Editor', () => {
         link_reach: 'public',
         link_role: 'editor',
         created_at: '2021-09-01T09:00:00Z',
+        title: '',
       });
 
-      await goToGridDoc(page);
+      const [randomDoc] = await createDoc(
+        page,
+        'doc-editor-ai',
+        browserName,
+        1,
+      );
 
-      await verifyDocName(page, 'Mocked document');
+      await verifyDocName(page, randomDoc);
 
       await page.locator('.bn-block-outer').last().fill('Hello World');