+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to
 
 ## Fixed
 
+- 🐛(frontend) conditionally render AI button only when feature is enabled #814
 - 🐛(backend) compute ancestor_links in get_abilities if needed #725
 - 🔒️(back) restrict access to document accesses #801